\docType{package}
\name{AntWeb}
\alias{AntWeb}
\alias{AntWeb-package}
\title{AntWeb}
\description{
AntWeb
}
\details{
\href{http://www.antweb.org/}{The AntWeb} world's largest
online database of images, specimen records, and natural
history information on ants. The database is maintained and
hosted by the [California Academy of
Sciences](http://www.calacademy.org/).
}

