% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aw_data_all}
\alias{aw_data_all}
\title{Download all aw_data available for any request}
\usage{
aw_data_all(..., progress = "text")
}
\arguments{
\item{...}{All the same arguments that get passed to \code{aw_data}}

\item{progress}{Default is on and set to \code{text}. Set to \code{none} to suppress}
}
\description{
This is a thin wrapper around aw_data
}
\examples{
\dontrun{
# crem <- aw_data_all(genus = "crematogaster", georeferenced = TRUE)
}
}
\seealso{
aw_data
}
\keyword{data}
\keyword{download}

