% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aw_distinct}
\alias{aw_distinct}
\title{aw_distinct}
\usage{
aw_distinct(rank = "genus", habitat = NULL, country = NULL,
  min_elevation = NULL, max_elevation = NULL, limit = 1000,
  offset = NULL)
}
\arguments{
\item{rank}{= "genus" Default is genus. But you can also use phylum, sub-phylum etc}

\item{habitat}{The habitat type}

\item{country}{Country name}

\item{min_elevation}{Min elevation recorded for specimen}

\item{max_elevation}{Max elevation recorded for specimen}

\item{limit}{= 1000 Default limit. Set higher if necessary}

\item{offset}{To be used in conjunction with limit}
}
\description{
Retrieves a data.frame of distinct ranks based on various restrictions
}
\examples{
\dontrun{
aw_distinct(rank = "genus", country = "Madagascar")
}
}

