% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_utils.R
\name{print.antweb}
\alias{print.antweb}
\title{Print a summary for an antweb object}
\usage{
\method{print}{antweb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{antweb}}

\item{...}{additional arguments}
}
\description{
Print a summary for an antweb object
}
