% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aw_code}
\alias{aw_code}
\title{aw_code}
\usage{
aw_code(occurrenceid = NULL, catalogNumber = NULL)
}
\arguments{
\item{occurrenceid}{A unique id in the AntWeb database identifying a particular specimen}

\item{catalogNumber}{Specimen catalogue number}
}
\value{
list
}
\description{
Retrieve data by specimen id
}
\examples{
# data_by_code <- aw_code(occurrenceid = "CAS:ANTWEB:alas188691")
# data_by_code <- aw_code(catalognumber="inb0003695883")
}
\seealso{
\code{\link{aw_data}}
}

