\name{aw_data_all}
\alias{aw_data_all}
\title{Download all aw_data available for any request}
\usage{
aw_data_all(...)
}
\arguments{
  \item{...}{All the same arguments that get passed to
  \code{aw_data}}
}
\description{
This is a thin wrapper around aw_data
}
\examples{
\dontrun{
# crem <- aw_data_all(genus = "crematogaster", georeferenced = TRUE)
}
}
\seealso{
aw_data
}
\keyword{data}
\keyword{download}

